# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "教室排课、教室资源管理和冲突检测 RESTful API。",
        "title": "教室排课助手 API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/teachers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "List teachers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Create a teacher",
                "parameters": [
                    {
                        "description": "teacher payload",
                        "name": "teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CreateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/teachers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get a teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "teacher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Update a teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "teacher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "teacher payload",
                        "name": "teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.UpdateTeacherRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Delete a teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "teacher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/time-slots": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-slots"
                ],
                "summary": "List time-slots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-slots"
                ],
                "summary": "Create a time slot",
                "parameters": [
                    {
                        "description": "time slot payload",
                        "name": "time_slot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CreateTimeSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/time-slots/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-slots"
                ],
                "summary": "Get a time slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "time slot id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-slots"
                ],
                "summary": "Update a time slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "time slot id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "time slot payload",
                        "name": "time_slot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.UpdateTimeSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-slots"
                ],
                "summary": "Delete a time slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "time slot id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_gbschedule_gbschedule_internal_dto.CreateTeacherRequest": {
            "type": "object",
            "required": [
                "employee_no",
                "name"
            ],
            "properties": {
                "contact": {
                    "type": "string",
                    "maxLength": 128
                },
                "employee_no": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unavailable_slots": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.CreateTimeSlotRequest": {
            "type": "object",
            "required": [
                "code",
                "end_time",
                "name",
                "start_time"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "end_time": {
                    "type": "string",
                    "maxLength": 8
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "start_time": {
                    "type": "string",
                    "maxLength": 8
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.UpdateTeacherRequest": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string",
                    "maxLength": 128
                },
                "employee_no": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unavailable_slots": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.UpdateTimeSlotRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "end_time": {
                    "type": "string",
                    "maxLength": 8
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "start_time": {
                    "type": "string",
                    "maxLength": 8
                }
            }
        }
    }
}